pyproject.toml
[tool.poetry]
name = "quoridorserver"
version = "0.1.0"
description = ""
authors = ["Tientjie <tientjieit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
fastapi = "^0.95.1"
pymongo = "^4.3.3"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
